+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Include summary of data streams in search responses. [#1264](https://github.com/elastic/package-registry/pull/1264)
+
 ### Deprecated
 
 ### Known Issues